dynamic programming
This presents another interesting problem: given an integer and the operations of +1, *2, and *3, we need to find the minimum number of operations needed to get to the given integer. We also need to print the integer after every operation, i.e., show how the answer is obtained, in what sequence. This last bit was fun to implement