recursive problem
This is by far one of the most challenging problems I have encountered. The problem is to divide the given array into three parts having the same sum, but not necessarily of the same size. The program says if it's possible to divide the array as such or not. The reference for this problem was obtained from the following link : https://youtu.be/qpgqhp_9d1s