Data structure problem
This program implements checking of code brackets in a string. It checks if all the opening brackets match all the closing brackets. The brackets allowed are '()' , '{}' and '[]'. In case the brackets match up, the programs prints Success, else it prints the index of the last unmatched closing bracket, and if there are no such closing brackets, it prints the index of the first unmatched opening bracket. This is a simplistic feature used in text editors to check code.