Data Structure Problem
This problem also replicates the simplistic feature of bracket matching of text editors, but rather than using inbuilt STL of stack, it uses a custom implementation of a generic stack. Implementing the stack helped better understand the features of C++ programming like class, structures and template class, and also shows the flexibility of structures. The stack is implemented using a linked list representation.